backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.39.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiohttp==3.9.1
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
